+++ week2/JS/sig_all/contants.js
@@ -1,0 +1,11 @@
+const form = document.querySelector('#login_form');
+const email = document.querySelector('#email');
+const inputs = document.querySelectorAll('.insert')
+const passwordInputs = document.querySelectorAll('.password');
+const password = document.querySelector('#password')
+const checkPassword = document.querySelector('#password_check')
+const btn = document.querySelector('.btn');
+const eye1 = document.querySelector('.eye1');
+const eye2 = document.querySelector('.eye2');
+
+export {email, inputs, password, passwordInputs, checkPassword, btn, eye1, eye2};
